Adjust NetID setting max value limit Now server send new NetID will increment from 1000

+++ SlayTheSpire2.LAN.Multiplayer/Patchs/NSettingsScreenPatch.cs
@@ -128,7 +128,7 @@ namespace SlayTheSpire2.LAN.Multiplayer.Patchs
 
                 netIdInput.Step = 1;
                 netIdInput.MinValue = 2;
-                netIdInput.MaxValue = double.MaxValue;
+                netIdInput.MaxValue = ulong.MaxValue;
                 netIdInput.Value = SettingsHelper.Instance.SettingsModel.NetId;
                 netIdInput.ValueChanged += value =>
                 {
